requirements.txt
Django==5.0.6
python-decouple==3.8
httpx==0.27.0
gunicorn==22.0.0
uvicorn==0.30.1

